pyproject.toml
[project]
name = "eew-client"
version = "0.1.0"
description = ""
authors = [
    {name = "mame7777",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "ruff (>=0.11.4,<0.12.0)",
    "requests (>=2.32.3,<3.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
]
packages = [
    { include = "eew_client", from = "src" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]
ignore = ["E501"]